+++ .gitignore
@@ -1,2 +1,8 @@
+.mvn/
 .idea/
+target/
+HELP.md
+car rental.iml
+mvnw
+mvnw.cmd
 demo.iml
