+++ src/main/java/com/svb/car/rental/controller/ReservationController.java
@@ -1,0 +1,16 @@
+package com.svb.car.rental.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+public class ReservationController {
+
+  String barcode="some_string";
+  @GetMapping("/reserve/{barcode}")
+  boolean reserveVehicle(@PathVariable String barcode) {
+    return false;
+  }
+}
