+++ src/main/java/com/svb/car/rental/service/HealthService.java
@@ -1,0 +1,9 @@
+package com.svb.car.rental.service;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Service;
+
+@Service
+public class HealthService {
+
+}
